- convert HttpConnectionUtils to Object class with AutoCloseable - move new class HttpConnectionWrapper out of util package
