+++ src/main/java/net/helpscout/api/ResultExtractor.java
@@ -1,9 +1,7 @@
 package net.helpscout.api;
 
-import java.net.HttpURLConnection;
-
 public interface ResultExtractor<T> {
 
-    T extract(HttpURLConnection conn);
+    T extract(HTTPConnectionWrapper conn);
 
 }
